# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Booking table application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories/{id}/uploadImage": {
            "post": {
                "tags": [
                    "categories"
                ],
                "summary": "Updates the category picture",
                "description": "Updates the category picture",
                "operationId": "categoryUploadImage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of category",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "Pass data to add a new category image",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "type": "file",
                                        "maxLength": 255,
                                        "example": "(file path)"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success storing a new user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "image_url": {
                                            "type": "string",
                                            "maxLength": 255,
                                            "example": "storage/CategoryImages/236095676.png"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "ModelNotFoundException handled for API"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/places/{id}/menu": {
            "get": {
                "tags": [
                    "places"
                ],
                "summary": "Places menu",
                "description": "Getting information about 1 institution",
                "operationId": "placesMenu",
                "responses": {
                    "200": {
                        "description": "Success getting information about 1 institution",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Coffee": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/places/{id}/reviewsCount": {
            "get": {
                "tags": [
                    "places"
                ],
                "summary": "Returns the number of reviews for a place",
                "description": "Returns the number of reviews for a place",
                "operationId": "reviewsCount",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of category",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success storing a new user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "reviews_count": {
                                            "type": "integer",
                                            "maxLength": 255,
                                            "example": 2
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "ModelNotFoundException handled for API"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "name": {
                        "description": "Product name",
                        "type": "string",
                        "example": "Americano"
                    },
                    "weight": {
                        "description": "Product weight",
                        "type": "string",
                        "example": "200ml"
                    },
                    "price": {
                        "description": "Product price",
                        "type": "number",
                        "example": 4
                    },
                    "image_url": {
                        "description": "On creation and update, accepts a file, but stores and returns a link to the file on the server",
                        "type": "string",
                        "example": "storage/products/coffe_americano.png"
                    },
                    "category_id": {
                        "description": "ID of the category the product belongs to",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "description": "Initial creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Product"
                }
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "api_key",
                "in": "header"
            }
        }
    }
}